+++ packages/api-core/src/site-plugins-scheduled-updates/fetchers.ts
@@ -1,10 +1,5 @@
 import { wpcom } from '../wpcom-fetcher';
-import type {
-	CreateScheduledUpdateBody,
-	EditScheduledUpdateBody,
-	MultisiteScheduledUpdatesResponse,
-	ScheduledUpdate,
-} from './types';
+import type { MultisiteScheduledUpdatesResponse, ScheduledUpdate } from './types';
 
 // Single-site list
 export async function fetchSiteScheduledUpdates(
@@ -12,42 +7,6 @@ export async function fetchSiteScheduledUpdates(
 ): Promise< Record< string, ScheduledUpdate > > {
 	return await wpcom.req.get( {
 		path: `/sites/${ siteId }/update-schedules`,
-		apiNamespace: 'wpcom/v2',
-	} );
-}
-
-// Single-site create
-export async function createSiteScheduledUpdate(
-	siteId: number,
-	body: CreateScheduledUpdateBody
-): Promise< unknown > {
-	return await wpcom.req.post( {
-		path: `/sites/${ siteId }/update-schedules`,
-		apiNamespace: 'wpcom/v2',
-		body,
-	} );
-}
-
-// Single-site edit
-export async function editSiteScheduledUpdate(
-	siteId: number,
-	scheduleId: string,
-	body: EditScheduledUpdateBody
-): Promise< unknown > {
-	return await wpcom.req.put( {
-		path: `/sites/${ siteId }/update-schedules/${ scheduleId }`,
-		apiNamespace: 'wpcom/v2',
-		body,
-	} );
-}
-
-// Single-site delete
-export async function deleteSiteScheduledUpdate(
-	siteId: number,
-	scheduleId: string
-): Promise< unknown > {
-	return await wpcom.req.delete( {
-		path: `/sites/${ siteId }/update-schedules/${ scheduleId }`,
 		apiNamespace: 'wpcom/v2',
 	} );
 }
